fix: refactor for code reuse
